reworked folder structure and files
