+++ .eslintrc.json
@@ -1,9 +1,10 @@
 {
   "env": {
     "browser": true,
-    "es2021": true
+    "es2021": true,
+    "jest": true
   },
-  "extends": ["airbnb", "plugin:prettier/recommended"],
+  "extends": ["airbnb", "plugin:react/recommended", "plugin:prettier/recommended"],
   "parserOptions": {
     "ecmaFeatures": {
       "jsx": true
@@ -13,13 +14,23 @@
   },
   "plugins": ["react"],
   "rules": {
-    "react/prop-types": "off",
     "prettier/prettier": [
       "error",
       {
         "endOfLine": "auto"
+      },
+      {
+        "usePrettierrc": true
       }
     ],
-    "react/react-in-jsx-scope": "off"
+    "react/react-in-jsx-scope": "off",
+    "react/jsx-props-no-spreading": [0],
+    "import/no-unresolved": [0],
+    "react/prop-types": [0],
+    "no-underscore-dangle": "off",
+    "react/no-danger": "off",
+    "no-restricted-exports": "off",
+    "react/display-name": "off",
+    "react/function-component-definition": [2, { "namedComponents": ["arrow-function", "function-declaration"] }]
   }
 }
