--- src/App/app.scss
@@ -1,12 +1,0 @@
-@import '/src/styles/globalStyles.scss';
-
-.loading-screen {
-  position: absolute;
-  background: $secondary-color;
-  z-index: 49;
-  pointer-events: none;
-
-  @media (max-width: $mobile-or-tablet-screen) {
-    z-index: 49;
-  }
-}
