+++ src/Views/About/About.scss
@@ -1,4 +1,6 @@
-@import '../../styles/globalStyles.scss';
+@use "/src/styles/mixins";
+@import "/src/styles/variables";
+@import "/src/styles/typography";
 
 .about-page {
   background: $secondary-color;
@@ -54,7 +56,7 @@
     }
   }
 
-  @media (max-width: $mobile-or-tablet-screen) {
+  @include mixins.respond(tab-land) {
     border-left: none;
     border-top: 2px solid $tertiary-color;
     height: calc(100vh - 64px);
