+++ src/Views/Home/Intro/Intro.scss
@@ -1,4 +1,6 @@
-@import '../../../styles/globalStyles.scss';
+@use "/src/styles/mixins";
+@import "/src/styles/variables";
+@import "/src/styles/typography";
 
 .sub-text {
   -webkit-animation: slide-in 2s cubic-bezier(0.215, 0.61, 0.355, 1) both;
@@ -124,10 +126,7 @@
         mix-blend-mode: normal;
         box-shadow: 3px 10px 20px 1px rgba(0, 0, 0, 0.616);
         text-shadow: 0px 3px 3px white;
-        background-image: radial-gradient(
-          rgb(53, 46, 46) 70%,
-          rgb(139, 125, 125)
-        );
+        background-image: radial-gradient(rgb(53, 46, 46) 70%, rgb(139, 125, 125));
         transform: scale(1.05);
         color: white;
         border: none;
@@ -135,7 +134,7 @@
     }
   }
 
-  @media (max-width: $mobile-or-tablet-screen) {
+  @include mixins.respond(tab-land) {
     font-size: 1.5rem;
     margin: 0;
     text-align: left;
@@ -185,10 +184,7 @@
           mix-blend-mode: normal;
           box-shadow: 3px 10px 20px 1px rgba(0, 0, 0, 0.616);
           text-shadow: 0px 3px 3px white;
-          background-image: radial-gradient(
-            rgb(53, 46, 46) 70%,
-            rgb(139, 125, 125)
-          );
+          background-image: radial-gradient(rgb(53, 46, 46) 70%, rgb(139, 125, 125));
           transform: scale(1.05);
           color: white;
           border: none;
