--- src/Views/Layout/HeaderMenu/HeaderMenu.scss
@@ -1,9 +1,0 @@
-@import '/src/styles/globalStyles.scss';
-
-.headerMenu {
-  padding: 0;
-
-  @media (max-width: $mobile-or-tablet-screen) {
-    padding: 0 2px;
-  }
-}
