+++ src/Views/Projects/ProjectSlider/ProjectSlider.scss
@@ -1,20 +1,21 @@
-@import '/src/styles/globalStyles.scss';
+@import "/src/styles/variables";
+@import "/src/styles/typography";
 
-.projectSlider {
+p .projectSlider {
   height: 100%;
   width: 100%;
 
-  div[class='ant-carousel'] {
+  div[class="ant-carousel"] {
     height: 100%;
     width: 100%;
   }
 
   &__sliderArrow {
-    &[id='prevArrow'] {
+    &[id="prevArrow"] {
       left: -3% !important;
     }
 
-    &[id='nextArrow'] {
+    &[id="nextArrow"] {
       right: -3% !important;
     }
   }
