+++ src/Views/Projects/Projects.scss
@@ -1,4 +1,6 @@
-@import '../../styles/globalStyles.scss';
+@use "/src/styles/mixins";
+@import "/src/styles/variables";
+@import "/src/styles/typography";
 
 .projects {
   position: absolute;
@@ -16,7 +18,7 @@
       margin: 0px;
       font-family: $secondary-font-family;
       font-weight: 900;
-      background: url('../../assets/header-pic.jpeg') 50% 50%;
+      background: url("../../assets/header-pic.jpeg") 50% 50%;
       background-size: cover;
       -webkit-text-fill-color: transparent;
       background-clip: text;
@@ -27,7 +29,7 @@
     height: 75%;
   }
 
-  @media (max-width: $mobile-or-tablet-screen) {
+  @include mixins.respond(tab-land) {
     &__header {
       &__text {
         font-size: 3rem;
