+++ src/components/ProjectCard/ProjectCard.scss
@@ -1,4 +1,6 @@
-@import '../../../styles/globalStyles.scss';
+@use "/src/styles/mixins";
+@import "/src/styles/variables";
+@import "/src/styles/typography";
 
 .card {
   width: 95%;
@@ -28,16 +30,13 @@
       justify-content: center;
       background-color: #ece9e6;
       text-align: center;
-      font-family: 'Monoton', cursive;
+      font-family: "Monoton", cursive;
       color: #2f2f2e;
       font-size: 1.5rem;
       font-weight: bolder;
-      box-shadow: -2px 2px 5px rgba(176, 176, 176, 0.2),
-        2px -2px 5px rgba(176, 176, 176, 0.2),
-        -2px -2px 5px rgba(216, 216, 216, 0.9),
-        2px 2px 13px rgba(176, 176, 176, 0.9),
-        inset 1px 1px 2px rgba(216, 216, 216, 0.3),
-        inset -1px -1px 2px rgba(176, 176, 176, 0.5);
+      box-shadow: -2px 2px 5px rgba(176, 176, 176, 0.2), 2px -2px 5px rgba(176, 176, 176, 0.2),
+        -2px -2px 5px rgba(216, 216, 216, 0.9), 2px 2px 13px rgba(176, 176, 176, 0.9),
+        inset 1px 1px 2px rgba(216, 216, 216, 0.3), inset -1px -1px 2px rgba(176, 176, 176, 0.5);
     }
 
     &__top__front {
@@ -76,7 +75,7 @@
     }
   }
 
-  @media only screen and (max-width: 700px) {
+  @include mixins.respond(tab-port) {
     min-height: 20rem;
     min-width: 80vw;
     margin: 10px 0px;
@@ -97,9 +96,3 @@
     }
   }
 }
-
-// @keyframes card-drop {
-//   to {
-//     transform: translateY(0);
-//   }
-// }
