--- src/index.scss
@@ -1,13 +1,0 @@
-body {
-    position: fixed;
-    height: 100%;
-    width: 100%;
-    margin: 0px;
-}
-
-#root {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    margin: 0px;
-}
