+++ src/layout/SideMenu/SideMenu.scss
@@ -1,9 +1,10 @@
-@import '/src/styles/globalStyles.scss';
+@use "src/styles/mixins";
+@import "/src/styles/variables";
 
 .menu {
   z-index: 50;
 
-  div[class='ant-layout-sider-children'] {
+  div[class="ant-layout-sider-children"] {
     display: flex;
     justify-content: center;
     align-items: center;
@@ -19,7 +20,7 @@
     align-items: flex-start;
   }
 
-  @media (max-width: $mobile-or-tablet-screen) {
+  @include mixins.respond(tab-land) {
     z-index: 50;
   }
 }
