+++ src/styles/_mixins.scss
@@ -1,0 +1,44 @@
+@mixin flex($horizonal, $vertical, $direction) {
+    display: flex;
+    flex-direction: $direction;
+    justify-content: $horizonal;
+    align-items: $vertical;
+  }
+  
+@mixin respond($breakpoint) {
+    @if $breakpoint==mini {
+      @media (max-width: 320px) {
+        @content;
+      }
+    }
+  
+    @if $breakpoint==phone {
+      @media (max-width: 600px) {
+        @content;
+      }
+    }
+  
+    @if $breakpoint==tab-port {
+      @media (max-width: 900px) {
+        @content;
+      }
+    }
+  
+    @if $breakpoint==tab-land {
+      @media (max-width: 1200px) {
+        @content;
+      }
+    }
+  
+    @if $breakpoint==desk-small {
+      @media (max-width: 1600px) {
+        @content;
+      }
+    }
+  
+    @if $breakpoint==desk-large {
+      @media (min-width: 1600px) {
+        @content;
+      }
+    }
+  }
