+++ src/styles/_variables.scss
@@ -1,0 +1,4 @@
+$primary-color: #886c57;
+$secondary-color: #413b3b;
+$tertiary-color: #ece9e6;
+$primary-bg-color: linear-gradient(to right, #ffffff, #ece9e6);
