--- src/styles/globalStyles.scss
@@ -1,12 +1,0 @@
-$primary-color: #886c57;
-$secondary-color: #413b3b;
-$tertiary-color: #ece9e6;
-$primary-bg-color: linear-gradient(to right, #ffffff, #ece9e6);
-$primary-font-family: 'Roboto', sans-serif;
-$secondary-font-family: 'Monoton', cursive;
-$nav-button-font-size: '';
-$text-primary-font-size: '';
-$text-secondary-font-size: '';
-$header-font-size: '';
-
-$mobile-or-tablet-screen: 1200px;
